Refactor user registration by moving it from AuthService to UserService
